+++ src/front/components/Navbar.jsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState, useRef, useCallback } from "react";
 import { Link, useNavigate } from "react-router-dom";
 import { ChefHat, Cog, DoorOpen, FilePlus, Heart, Bookmark, MessageCircle, Menu, User } from "lucide-react";
 import UserAvatar from "./UserAvatar";
-import socketService from "../utils/socketService";
+import socketService from "../shared/utils/socketService";
 
 import ThemeToggle from "./ThemeToggle"; //Imported for final CSS styling implementation
