+++ src/front/main.jsx
@@ -5,11 +5,11 @@ import ReactDOM from 'react-dom/client'
 // CUSTOM CSS AND SCSS
 
 // SCSS imports
-import './styles/scss-bootstrap-custom-theme.scss'; // COMPLETE Custom Bootstrap theme with SCSS variables
-// import './styles/simple-custom.scss';            // SIMPLE   Custom Bootstrap theme with SCSS variables
+import '../shared/styles/scss-bootstrap-custom-theme.scss'; // COMPLETE Custom Bootstrap theme with SCSS variables
+// import '../shared/styles/simple-custom.scss';            // SIMPLE   Custom Bootstrap theme with SCSS variables
 
 // CSS custom theme 
-// import './styles/custom-theme.css'
+// import '../shared/styles/custom-theme.css'
 // ----------------------------------------------------------------------
