+++ src/front/messages/components/Websocket-Status/WebSocketStatus.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import socketService from '../../../utils/socketService';
+import socketService from '../../../shared/utils/socketService';
 
 /**
  * WebSocket Status Component
