+++ src/front/pages/Home.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from "react"
 import { Link } from "react-router-dom";
 import useGlobalReducer from "../hooks/useGlobalReducer.jsx";
-import food from "../assets/img/table_food.jpg";
+import food from "../shared/assets/img/table_food.jpg";
 
 export const Home = () => {
