+++ src/front/shared/utils/socketService.js
